FEM: Python base constraint, split in obj and gui module
