+++ src/Mod/Fem/PyGui/_ViewProviderFemConstraintFlowVelocity.py
@@ -28,14 +28,14 @@ __url__ = "http://www.freecadweb.org"
 
 import FreeCAD as App
 import FemUtils
-import FemConstraint
+import ViewProviderFemConstraint
 from FreeCAD import Units
 
 import FreeCADGui as Gui
 from . import FemSelectionWidgets
 
 
-class ViewProxy(FemConstraint.ViewProxy):
+class ViewProxy(ViewProviderFemConstraint.ViewProxy):
 
     def getIcon(self):
         return ":/icons/fem-constraint-flow-velocity.svg"
